Most important Zork++ abstractions (traits) deserves its own place, decoupled from domain
