+++ zork++/src/lib/lib.rs
@@ -1,3 +1,4 @@
+pub mod bounds;
 pub mod cli;
 pub mod compiler;
 pub mod config_file;
