+++ zork++/src/lib/project_model/mod.rs
@@ -7,14 +7,12 @@ pub mod project;
 pub mod sourceset;
 pub mod tests;
 
-use std::{
-    fmt::{Debug, Display},
-    path::Path,
-};
+
+use std::fmt::Debug;
 
 use self::{
-    arguments::Argument, build::BuildModel, compiler::CompilerModel, executable::ExecutableModel,
-    modules::ModulesModel, project::ProjectModel, sourceset::SourceSet, tests::TestsModel,
+    build::BuildModel, compiler::CompilerModel, executable::ExecutableModel,
+    modules::ModulesModel, project::ProjectModel, tests::TestsModel,
 };
 
 #[derive(Debug, PartialEq, Eq)]
@@ -25,36 +23,4 @@ pub struct ZorkModel<'a> {
     pub executable: ExecutableModel<'a>,
     pub modules: ModulesModel<'a>,
     pub tests: TestsModel<'a>,
-}
-
-pub trait ExtraArgs<'a> {
-    fn extra_args(&'a self) -> &'a [Argument<'a>];
-}
-
-pub trait ExecutableTarget<'a>: ExtraArgs<'a> {
-    fn name(&'a self) -> &'a str;
-    fn sourceset(&'a self) -> &'a SourceSet<'a>;
-}
-
-/// Represents any kind of translation unit and the generic operations
-/// applicable to all the implementors
-pub trait TranslationUnit: Display + Debug {
-    /// Outputs the declared filename for `self` being the translation unit
-    fn filename(&self) -> &Path;
-
-    fn filestem(&self) -> &str {
-        self.filename().file_stem().unwrap().to_str().unwrap()
-    }
-}
-
-impl TranslationUnit for &str {
-    fn filename(&self) -> &Path {
-        Path::new(self)
-    }
-}
-
-impl TranslationUnit for String {
-    fn filename(&self) -> &Path {
-        Path::new(self)
-    }
 }
